Add instance ID to metric labels
This allows to identify different Collector instances indenpently of how it is deployed.
