+++ obsreport/obsreport_receiver.go
@@ -240,6 +240,7 @@ func ReceiverContext(
 		mutators := []tag.Mutator{
 			tag.Upsert(tagKeyReceiver, receiver, tag.WithTTL(tag.TTLNoPropagation)),
 			tag.Upsert(tagKeyTransport, transport, tag.WithTTL(tag.TTLNoPropagation)),
+			tag.Upsert(tagKeyInstanceID, instanceID, tag.WithTTL(tag.TTLNoPropagation)),
 		}
 		ctx, _ = tag.New(ctx, mutators...)
 	}
@@ -281,6 +282,30 @@ func traceReceiveOp(
 	span.AddAttributes(trace.StringAttribute(
 		TransportKey, transport))
 	return ctx
+}
+
+// setParentLink tries to retrieve a span from parentCtx and if one exists
+// sets its SpanID, TraceID as a link to the given child Span.
+// It returns true only if it retrieved a parent span from the context.
+//
+// This is typically used when the parentCtx may already have a trace and is
+// long lived (eg.: an gRPC stream, or TCP connection) and one desires distinct
+// traces for individual operations under the long lived trace associated to
+// the parentCtx. This function is a helper that encapsulates the work of
+// linking the short lived trace/span to the longer one.
+func setParentLink(parentCtx context.Context, childSpan *trace.Span) bool {
+	parentSpanFromRPC := trace.FromContext(parentCtx)
+	if parentSpanFromRPC == nil {
+		return false
+	}
+
+	psc := parentSpanFromRPC.SpanContext()
+	childSpan.AddLink(trace.Link{
+		SpanID:  psc.SpanID,
+		TraceID: psc.TraceID,
+		Type:    trace.LinkTypeParent,
+	})
+	return true
 }
 
 // endReceiveOp records the observability signals at the end of an operation.
